requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
google-generativeai==0.4.0
python-dotenv==1.0.0
pydantic==2.4.2
